[FEAT/#33] 전체적인 UI 정리
